# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfferCreate": {
        "properties": {
          "category_l1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category L1"
          },
          "category_l2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category L2"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "geo_scope": {
            "items": {
              "type": "string"
            },
            "title": "Geo Scope",
            "type": "array"
          },
          "intent_type": {
            "enum": [
              "find_supplier",
              "find_buyer",
              "find_distributor",
              "find_local_partner",
              "find_manufacturer",
              "co_marketing",
              "find_investment",
              "service_partnership"
            ],
            "title": "Intent Type",
            "type": "string"
          },
          "structured_attrs": {
            "additionalProperties": true,
            "title": "Structured Attrs",
            "type": "object"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "intent_type",
          "title"
        ],
        "title": "OfferCreate",
        "type": "object"
      },
      "OfferResponse": {
        "properties": {
          "business_id": {
            "format": "uuid",
            "title": "Business Id",
            "type": "string"
          },
          "category_l1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category L1"
          },
          "category_l2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category L2"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "geo_scope": {
            "items": {
              "type": "string"
            },
            "title": "Geo Scope",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intent_type": {
            "enum": [
              "find_supplier",
              "find_buyer",
              "find_distributor",
              "find_local_partner",
              "find_manufacturer",
              "co_marketing",
              "find_investment",
              "service_partnership"
            ],
            "title": "Intent Type",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "structured_attrs": {
            "additionalProperties": true,
            "title": "Structured Attrs",
            "type": "object"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "intent_type",
          "title",
          "id",
          "business_id",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "OfferResponse",
        "type": "object"
      },
      "OfferUpdate": {
        "properties": {
          "category_l1": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category L1"
          },
          "category_l2": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category L2"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "geo_scope": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Scope"
          },
          "intent_type": {
            "anyOf": [
              {
                "enum": [
                  "find_supplier",
                  "find_buyer",
                  "find_distributor",
                  "find_local_partner",
                  "find_manufacturer",
                  "co_marketing",
                  "find_investment",
                  "service_partnership"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Type"
          },
          "structured_attrs": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Attrs"
          },
          "title": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "OfferUpdate",
        "type": "object"
      },
      "PersonCreate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "owner",
                  "director",
                  "sales_rep",
                  "authorized_rep"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "zalo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zalo Id"
          }
        },
        "required": [
          "full_name"
        ],
        "title": "PersonCreate",
        "type": "object"
      },
      "PersonResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "owner",
                  "director",
                  "sales_rep",
                  "authorized_rep"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "zalo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zalo Id"
          }
        },
        "required": [
          "full_name",
          "id",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "PersonResponse",
        "type": "object"
      },
      "PersonUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "owner",
                  "director",
                  "sales_rep",
                  "authorized_rep"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "role_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title"
          },
          "zalo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zalo Id"
          }
        },
        "title": "PersonUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Linko API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/businesses/{business_id}/offers": {
      "get": {
        "operationId": "list_offers_api_v1_businesses__business_id__offers_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OfferResponse"
                  },
                  "title": "Response List Offers Api V1 Businesses  Business Id  Offers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Offers",
        "tags": [
          "offers"
        ]
      },
      "post": {
        "operationId": "create_offer_api_v1_businesses__business_id__offers_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Offer",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/persons": {
      "get": {
        "operationId": "list_persons_api_v1_businesses__business_id__persons_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonResponse"
                  },
                  "title": "Response List Persons Api V1 Businesses  Business Id  Persons Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Persons",
        "tags": [
          "persons"
        ]
      },
      "post": {
        "operationId": "create_person_api_v1_businesses__business_id__persons_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Business Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Person",
        "tags": [
          "persons"
        ]
      }
    },
...
    "/api/v1/offers/{offer_id}": {
      "delete": {
        "operationId": "delete_offer_api_v1_offers__offer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Offer",
        "tags": [
          "offers"
        ]
      },
      "patch": {
        "operationId": "update_offer_api_v1_offers__offer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "offer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Offer",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/v1/persons/{person_id}": {
      "delete": {
        "operationId": "delete_person_api_v1_persons__person_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Person",
        "tags": [
          "persons"
        ]
      },
      "patch": {
        "operationId": "update_person_api_v1_persons__person_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Person",
        "tags": [
          "persons"
        ]
      }
    },
...
  }
}